Sprint 3

See merge request cse1105/2022-2023/teams/oopp-team-45!57

+++ client/src/main/java/client/scenes/MainSceneCtrl.java
@@ -1,6 +1,5 @@
 package client.scenes;
 
-
 import client.utils.ServerUtils;
 import com.google.inject.Inject;
 import commons.Board;
@@ -9,21 +8,28 @@ import commons.TaskList;
 import javafx.collections.FXCollections;
 import javafx.collections.ObservableList;
 import javafx.fxml.FXML;
+import javafx.scene.control.Alert;
 import javafx.scene.control.Button;
 import javafx.scene.control.Label;
 import javafx.scene.control.ListView;
 
 import javafx.scene.input.Clipboard;
 import javafx.scene.input.ClipboardContent;
+import javafx.stage.Modality;
 
 import java.io.IOException;
+import java.util.ArrayList;
+import java.util.List;
 
 public class MainSceneCtrl {
 
-    private final ServerUtils server;
+    private  ServerUtils server;
     private final MainCtrlTalio mainCtrl;
     private final RenameCtrl renameCtrl;
 
+    private final long DEFAULT_ID;
+
+    List<TaskListCtrl> taskListCtrls;
 
     ObservableList<TaskList> listData;
 
@@ -46,6 +52,7 @@ public class MainSceneCtrl {
 
     /**
      * constructor
+     *
      * @param mainCtrl the main controller
      */
     @Inject
@@ -54,24 +61,42 @@ public class MainSceneCtrl {
         this.server = server;
         this.mainCtrl = mainCtrl;
         this.renameCtrl = renameCtrl;
+
+        this.DEFAULT_ID = server.getDefaultId();
     }
 
     /**
      * initialize the scene with the listview elements as the TaskList scene
      */
-    public void initialize() {
+    public void initialize(ServerUtils server) {
+        this.server = server;
+
+        taskListCtrls = new ArrayList<>();
+
+        if (mainCtrl.getActiveBoard() == null) {
+            mainCtrl.setActiveBoard(server.getDefaultBoard());
+        }
+
         listData = FXCollections.observableArrayList();
+
+        if (lists == null) {
+            lists = new ListView<>();
+        }
+
         lists.setFixedCellSize(0);
         lists.setItems(listData);
-        lists.setCellFactory(new TaskListCtrl(server, this, mainCtrl,
-                renameCtrl));
+        lists.setCellFactory(taskListView -> new TaskListCell(new TaskListCtrl(
+                server, this, mainCtrl, renameCtrl), this));
         refresh();
     }
 
     /**
      * go back to the connect screen
+     * TODO: delete all the potential local storage,
+     * since the user want to connect to a different server
      */
     public void back() {
+        ServerUtils.resetServer();
         mainCtrl.showConnect();
     }
 
@@ -79,10 +104,16 @@ public class MainSceneCtrl {
      * refresh the list
      */
     public void refresh() {
-        listData = FXCollections.observableList(server.getTaskList());
+//        listData = FXCollections.observableList(server.getTaskList());
+        listData = FXCollections.observableList(
+                server.getBoardData(mainCtrl.getActiveBoard().getId()));
         taskData = FXCollections.observableList(server.getTasks());
         lists.setItems(listData);
         tasks.setItems(taskData);
+
+        for (TaskListCtrl taskListCtrl : taskListCtrls) {
+            taskListCtrl.refresh();
+        }
     }
 
     private int i = 0;
@@ -98,6 +129,18 @@ public class MainSceneCtrl {
      * Rename the current board
      */
     public void renameBoard() {
+        Board board = mainCtrl.getActiveBoard();
+        if (board == null) {
+            System.out.println("Cannot rename board: this is a dummy board!");
+            return;
+        }
+        if (board.getId() == DEFAULT_ID) {
+            var alert = new Alert(Alert.AlertType.ERROR);
+            alert.initModality(Modality.APPLICATION_MODAL);
+            alert.setContentText("You cannot rename the default board!");
+            alert.showAndWait();
+            return;
+        }
         mainCtrl.showRenameBoard();
     }
 
@@ -112,9 +155,16 @@ public class MainSceneCtrl {
             System.out.println("Cannot delete board: this is a dummy board!");
             return;
         }
-        mainCtrl.setActiveBoard(null);
+        if (board.getId() == DEFAULT_ID) {
+            var alert = new Alert(Alert.AlertType.ERROR);
+            alert.initModality(Modality.APPLICATION_MODAL);
+            alert.setContentText("You cannot delete the default board!");
+            alert.showAndWait();
+            return;
+        }
+        mainCtrl.setActiveBoard(server.getDefaultBoard());
         server.deleteBoard(board);
-        mainCtrl.showConnect();
+        refresh();
     }
 
     /**
@@ -129,8 +179,7 @@ public class MainSceneCtrl {
         Board board = mainCtrl.getActiveBoard();
         if (board == null) {
             System.out.println("This is the default board!");
-        }
-        else {
+        } else {
             String code = board.getCode();
             content.putString(code);
             clipboard.setContent(content);
@@ -155,12 +204,12 @@ public class MainSceneCtrl {
     }
 
     /**
-     * Edit a task
-     * @throws IOException
+     * View the details of a task
+     * @throws IOException -
      */
-    public void editTask() throws IOException {
+    public void viewTask() throws IOException {
         Task currentTask = tasks.getSelectionModel().getSelectedItem();
-        mainCtrl.showEditTask(currentTask);
+        mainCtrl.showTaskDetails(currentTask);
     }
 
 }
